Devirtualize LoadFeeTrack
