+++ Builds/VisualStudio2013/RippleD.vcxproj.filters
@@ -2799,12 +2799,9 @@
     <ClCompile Include="..\..\src\ripple\core\impl\LoadEvent.cpp">
       <Filter>ripple\core\impl</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\ripple\core\impl\LoadFeeTrackImp.cpp">
+    <ClCompile Include="..\..\src\ripple\core\impl\LoadFeeTrack.cpp">
       <Filter>ripple\core\impl</Filter>
     </ClCompile>
-    <ClInclude Include="..\..\src\ripple\core\impl\LoadFeeTrackImp.h">
-      <Filter>ripple\core\impl</Filter>
-    </ClInclude>
     <ClCompile Include="..\..\src\ripple\core\impl\LoadMonitor.cpp">
       <Filter>ripple\core\impl</Filter>
     </ClCompile>
