+++ src/ripple/core/tests/Config.test.cpp
@@ -18,7 +18,7 @@
 //==============================================================================
 
 #include <BeastConfig.h>
-#include <ripple/core/impl/LoadFeeTrackImp.h>
+#include <ripple/core/LoadFeeTrack.h>
 #include <ripple/core/Config.h>
 #include <ripple/core/ConfigSections.h>
 #include <ripple/basics/TestSuite.h>
