+++ src/ripple/core/tests/LoadFeeTrack.test.cpp
@@ -18,7 +18,7 @@
 //==============================================================================
 
 #include <BeastConfig.h>
-#include <ripple/core/impl/LoadFeeTrackImp.h>
+#include <ripple/core/LoadFeeTrack.h>
 #include <ripple/core/Config.h>
 #include <beast/unit_test/suite.h>
 
@@ -30,7 +30,7 @@ public:
     void run ()
     {
         Config d; // get a default configuration object
-        LoadFeeTrackImp l;
+        LoadFeeTrack l;
 
         expect (l.scaleFeeBase (10000, d.FEE_DEFAULT, d.TRANSACTION_FEE_BASE) == 10000);
         expect (l.scaleFeeLoad (10000, d.FEE_DEFAULT, d.TRANSACTION_FEE_BASE, false) == 10000);
