+++ src/ripple/unity/core.cpp
@@ -21,7 +21,7 @@
 
 #include <ripple/core/impl/Config.cpp>
 #include <ripple/core/impl/DatabaseCon.cpp>
-#include <ripple/core/impl/LoadFeeTrackImp.cpp>
+#include <ripple/core/impl/LoadFeeTrack.cpp>
 #include <ripple/core/impl/LoadEvent.cpp>
 #include <ripple/core/impl/LoadMonitor.cpp>
 #include <ripple/core/impl/Job.cpp>
